requirements.txt
numpy
torch
gymnasium[atari]
ale-py
autorom[accept-rom-license]
pyyaml
opencv-python
